ChangedIdType

+++ src/main/java/com/labC/libraryCheckoutSystem/BookshelfRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public class BookshelfRepo extends CrudRepository<Bookshelf, Integer> {
+public class BookshelfRepo extends CrudRepository<Bookshelf, Long> {
 }
